AI API/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0
chromadb==0.4.22
sentence-transformers==2.5.1
pymssql
typing-extensions==4.9.0
numpy==1.26.4
mistralai==0.0.12
scikit-learn==1.4.1.post1
tqdm==4.66.1
torch==2.2.1
python-dateutil==2.8.2
pydantic-settings==2.1.0
aiohttp==3.9.3
asyncio==3.4.3
PyJWT==2.8.0
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
PyPDF2
huggingface_hub[hf_xet]
